+++ Changelog.md
@@ -1,5 +1,12 @@
 # graph-explorer Change Log
 
+## Upcoming
+
+The next release will include the following feature enhancements and bug fixes:
+
+**Features**
+- Added Default Connection support (https://github.com/aws/graph-explorer/pull/108)
+
 ## Release 1.3.1
 
 This patch release includes bugfixes for Release 1.3.0.